add auth for crete,update,delete

+++ main.py
@@ -58,10 +58,15 @@ def read_post(post_id: int, db: Session = Depends(get_db)):
 
 # Update a post by ID (only accessible to the author of the post)
 @app.put("/posts/{post_id}", response_model=schema.Post)
-def update_post(post_id: int, post: schema.PostCreate, db: Session = Depends(get_db)):
+def update_post(post_id: int, post: schema.PostCreate, db: Session = Depends(get_db), current_user: models.User = Depends(get_current_user)):
+    db_post = crud.get_post(db=db, post_id=post_id)
+    if db_post.owner_id != current_user.id and current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Insufficient permissions")
     return crud.update_post(db=db, post_id=post_id, post_data=post)
 
 # Delete a post by ID(only accessible to admin)
 @app.delete("/posts/{post_id}")
-def delete_post(post_id: int, db: Session = Depends(get_db)):
+def delete_post(post_id: int, db: Session = Depends(get_db), current_user: models.User = Depends(get_current_user)):
+    if current_user.role != "admin":
+        raise HTTPException(status_code=403, detail="Admin access required")
     return crud.delete_post(db=db, post_id=post_id)
